Fix export.go:355: cannot use size (type int64) as type float64 in argument to units.HumanSize

+++ export.go
@@ -352,7 +352,7 @@ func (e *Export) PrintHistory() {
 
 		debug("  - ", order[i].LayerConfig.Id[0:12],
 			humanDuration(time.Now().UTC().Sub(order[i].LayerConfig.Created.UTC())),
-			cmd, units.HumanSize(size))
+			cmd, units.HumanSize(float64(size)))
 	}
 }
